Revert "fixed lod, changed chunk data to use ints instead"
This reverts commit a887ca428458e679d06a4b40d4da7599439e79ea.

+++ Assets/Universe/Data/Generation/SeedBasedTerrainGenerator.cs
@@ -60,7 +60,7 @@ namespace Universe.Data.Generation {
 						int worldZ = worldOffsetZ + z;
 
 						// Block index within chunk
-						int blockIndex = chunkData.GetBlockIndex(new Vector3(x, y, z));
+						long blockIndex = chunkData.GetBlockIndex(new Vector3(x, y, z));
 
 						// Generate block type using asteroid generation logic
 						short blockType = GenerateAsteroidBlock(
